🔨 Rename single word components
