+++ frontend/src/modals/SongDelete.vue
@@ -1,5 +1,5 @@
 <template>
-	<modal :active="active" :title="t('modal.deleteSong')" @closed="emit('closed')">
+	<modal-dialog :active="active" :title="t('modal.deleteSong')" @closed="emit('closed')">
 		<div class="flex flex-col gap-2">
 			<div>{{ t('text.reallyDeleteSong', { title: title }) }}</div>
 			<div class="text-rose-600">{{ t('text.cannotBeUndone') }}</div>
@@ -22,7 +22,7 @@
 				</primary-button>
 			</div>
 		</div>
-	</modal>
+	</modal-dialog>
 </template>
 
 <script setup>
@@ -32,7 +32,7 @@ import { throwError } from '@/utils.js';
 import { useI18n } from 'vue-i18n';
 import { useRoute, useRouter } from 'vue-router'
 import { updateDoc, deleteDoc, doc } from 'firebase/firestore';
-import Modal from '@/elements/Modal.vue';
+import ModalDialog from '@/elements/ModalDialog.vue';
 import PrimaryButton from '@/elements/PrimaryButton.vue';
 
 // icons
